CSVParser

Description

Module provides functions for parsing between strings, sqlValue, Day and Row data type

Documentation

data Row

Custom data type to which rows from CSV are parsed

Constructors

Row

Fields

date :: Day

high :: Double

low :: Double

rowToSql :: Row -> [SqlValue]

Parses Row custom data type into array of SQL values containing the same data

formatSqlRow :: [SqlValue] -> String

Formats rows fetched from SQL for printing.

stringToDate :: String -> Day

Converts date data format from CSV into Day type

createRow :: [String] -> Row

Parses array of strings into object of type Row

csvToRows :: String -> [Row]

Converts CSV in string format into array of Rows