DBConnector

Description

Module provides connectivity functions with MySQL database

Documentation

connectToMySql :: IO Connection

Returns connection to university MySQL database

createDB :: IO ()

Creates tables in the stocks.db database

dropTables :: IO ()

Drops the tables stocks and companies from stocks.db database

storeRows :: String -> [Row] -> IO ()

Takes company name and rows and inserts company and stock prices into database, ignoring existing values

printHeader :: IO ()

Function for printing header, used by printHighest and printLowest functions

printHighest :: IO ()

Queries and prints the maximum value of column high for each company

printLowest :: IO ()

Queries and prints the minimum value of column low for each company

getCompanies :: IO [String]

Returns names of companies from the database